Report Ruby environment metadata
Collect environment metadata about which Ruby version (1.9.3/2.7.1),
engine (MRI/JRuby/Rubinius) and engine version is loaded.

Move the spec helpers to capture calls to the extension's
`set_environment_metadata` to a module that can be included separately
in every spec that wants to assert calls to this method. We use this
method of capturing method calls because the extension does not have a
"getter" for the environment metadata.
